refactor: move KV cache helpers
